Handle statements without payment details
Important Note: This patch changes the naming scheme!

Not all items will have a `payDetailUri`. Specifically, if the pay statement
is an adjustment (ie. raise) or reimbursement, there will exist only a
`statementImageUri`. This patch handles the scenario by using the last path
component of the `statementImageUri` as the `id` instead of the last path
component of `payDetailUri`. The later appears to always be an integer while
the former is a string whose first three or so characters appear to identify
the company. Because this string identifier is always present (while the
integer version is absent when `payDetailUri` is absent), I think this is
probably the most appropriate for use in the file name.