Bump 0.2.0
This version changes the way statements are organized to the following
structure:

    yyyy/
      yyyy-ID.json
      yyyy-ID.pdf
      yyyy-mm-dd-ID.json
      yyyy-mm-dd-ID.pdf

where `yyyy`, `mm`, and `dd` refer to the years, months and days of
those statements.

It also uses a different `ID` scheme, since the PDF ID is always
available and the JSON ID is not necessary always present.

You might want to backup your existing statements before upgrading the
gem and manually merge older statements that might not be available
through the API anymore.

Contributors:
- @cfilipov